+++ integrations/slack/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @trigger.dev/slack
 
+## 3.0.0-beta.25
+
+### Patch Changes
+
+- @trigger.dev/sdk@3.0.0-beta.25
+
+## 3.0.0-beta.24
+
+### Patch Changes
+
+- @trigger.dev/sdk@3.0.0-beta.24
+
 ## 3.0.0-beta.23
 
 ### Patch Changes