+++ packages/core-backend/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @trigger.dev/core-backend
 
+## 3.0.0-beta.25
+
+## 3.0.0-beta.24
+
 ## 3.0.0-beta.23
 
 ## 3.0.0-beta.22